test: stop keeping three copies of the same test scaffolding
`api-db` had its own copy of `setup_test_logging`, and the comment above it said so out loud -- *"This duplicates code from api-test-helper, but we don't want to take a dependency on that. Copy/pasting is fine."* Fair enough at the time, but `carbide-test-support` is a leaf crate with nothing in it but `tracing` and `tracing-subscriber`, so there's no dependency to be shy about. The two real implementations were identical apart from the crate name in the panic message, which is now a parameter.

The four `dpf/` suites had each grown their own `dpf_config` and `get_host_state`. Byte-identical, all of them -- I hashed them before touching anything, and the hoisted copy hashes the same as all four originals.

Primary callouts are:

- `setup_test_logging` lives in `carbide-test-support` now and takes the component name. `api-core` needed `carbide-test-support` promoted from a dev-dependency to an optional one behind its existing `test-support` feature, because `api-core::test_support` compiles into the *lib* when `carbide-api-web` turns that feature on -- a dev-dependency never reaches it. Both build paths verified, and `check-workspace-deps` is happy.
- `dpf_config` and `get_host_state` move up into `dpf/mod.rs`; seven definitions become two.
- `expected_switch.rs` gives up **176 lines** of `field: None` across 29 `ExpectedSwitch` literals. It's a `::prost::Message`, so every field dropped defaults to exactly what was written there -- `Option` to `None`, `String` to empty. The file already used `..Default::default()` twice; now it does so everywhere.

Coverage A/B over `api-core`, `api-db` and `test-support`, baseline against the same `06495cda5` this branch sits on: 2,020 tests pass on both sides, and production coverage comes out flat -- 12,543 functions covered before, 12,542 after, with `carbide_test_support::setup_test_logging` newly covered as you'd expect. The one function on the wrong side of that arithmetic is `machine_controller::handler::is_host_powered_off`, which this change cannot reach; it sits behind a race -- the host is only checked while a DPU hasn't yet reported UP -- and its execution count has measured 8, 5, 7, 4, 1 and 0 across six runs on three different commits. It happened to land on 1 in the baseline and 0 here.

One thing I went looking for and left alone: `api-test-helper/src/mock_rms.rs` is 1,525 lines and the obvious next candidate, since `carbide_rack::rms_client::test_support` implements the same `RmsApi` trait. It isn't a swap. `mock_rms.rs` is macro-generated -- `impl_enqueue_inspect!` emits the `enqueue_*`/`*_calls` pairs -- and `component-manager` leans on 22 of those generated methods across 63 call sites, none of which `RmsSim` has. That's a rewrite of 63 test call sites onto an API built along entirely different lines for no change in what's covered, so it stays where it is.

The site-explorer report builders are still outstanding too; those want designing rather than sweeping.

Tests updated!

This supports https://github.com/NVIDIA/infra-controller/issues/4225
