+++ crates/api-core/src/tests/dpf/mod.rs
@@ -19,3 +19,32 @@ mod happy_path;
 mod reprovisioning;
 mod stale_labels;
 mod waiting_for_ready;
+
+use model::machine::ManagedHostState;
+
+use crate::tests::common::api_fixtures::TestEnv;
+use crate::tests::common::api_fixtures::test_managed_host::TestManagedHost;
+
+// The DPF suites all want the same starting point: DPF on, with a BF3 BFB URL to hand out.
+// Each of the four files below had grown its own byte-identical copy of this.
+fn dpf_config() -> crate::cfg::file::DpfConfig {
+    crate::cfg::file::DpfConfig {
+        enabled: true,
+        deployments: crate::cfg::file::DpfDeploymentsConfig {
+            bf3: crate::cfg::file::DpfDeploymentConfig {
+                bfb_url: Some("http://example.com/test.bfb".to_string()),
+                ..Default::default()
+            },
+            ..Default::default()
+        },
+        ..Default::default()
+    }
+}
+
+// Reads the host's committed state straight from the database rather than through the RPC
+// surface, which is what these suites assert transitions against.
+async fn get_host_state(env: &TestEnv, mh: &TestManagedHost) -> ManagedHostState {
+    let mut txn = env.db_txn().await;
+    let machine = mh.host().db_machine(&mut txn).await;
+    machine.state.value
+}
